+++ packages/ciphernode/core/src/serializers/plaintext.rs
